requirements.txt
albumentations==2.0.8
gradio==6.11.0
numpy==2.4.4
opencv_python==4.13.0.92
PyYAML==6.0.3
torch==2.6.0+cu124
ultralytics==8.4.33
